pMetaload/requirements.txt
Flask==1.0
requests==2.18.4
pytest==3.5.1
beautifulsoup4==4.6.0
GitPython==2.1.9
selenium
dottorrent